python/mcp-test/pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "fastmcp>=2.8.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.1",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
]
